fix: address review comments for Hytte-818y
- Move UnlockedAchievement type to shared lib/regnemester/achievements/types.ts, eliminating the upward dep from feedback/ into components/
- Suppress finish fanfare/milestone sound in Marathon and Blitz when achievements were unlocked (overlay plays its own fanfare per unlock, avoiding duplicates)
- Add legacy 'Spacebar' key value to overlay dismiss handler for cross-browser compat
- Fix queue-hint test assertion to match stableT mock output (key string, not translated text)
