+++ web/src/lib/regnemester/achievements/types.ts
@@ -1,0 +1,7 @@
+export interface UnlockedAchievement {
+  code: string
+  title: string
+  description: string
+  tier: string
+  unlocked_at: string
+}
